Move the character set definitions into a separate header file.
