+++ src/terminal/adapter/lib/adapter.vcxproj
@@ -24,6 +24,7 @@
   <ItemGroup>
     <ClInclude Include="..\adaptDefaults.hpp" />
     <ClInclude Include="..\adaptDispatch.hpp" />
+    <ClInclude Include="..\charsets.hpp" />
     <ClInclude Include="..\DispatchTypes.hpp" />
     <ClInclude Include="..\DispatchCommon.hpp" />
     <ClInclude Include="..\InteractDispatch.hpp" />
