+++ src/terminal/adapter/lib/adapter.vcxproj.filters
@@ -80,6 +80,9 @@
     <ClInclude Include="..\termDispatch.hpp">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\charsets.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Natvis Include="$(SolutionDir)tools\ConsoleTypes.natvis" />
